Remove `api.ts` and all related AI chat service functions
Simplifies the codebase by deleting `api.ts`, including all associated API functions such as AI chat requests, feedback submission, actions planning, translation handling, and related utilities. Prepares for a transition to a new architecture or integration approach.
